[Refactor] Do not rely on implicit conversions in {fmt} calls  (#72022)
(cherry picked from commit bbef286d9d80969f3c9068cd98a04ba62090ed3c)

# Conflicts:
#	be/src/exec/pipeline/exchange/sink_buffer.cpp
#	be/src/exec/pipeline/fetch_processor.cpp
#	be/src/util/compression/stream_compressor.cpp

+++ be/src/util/compression/stream_compression.cpp
@@ -1479,7 +1479,7 @@ Status StreamCompression::create_decompressor(CompressionTypePB type,
         *decompressor = std::make_unique<LzoStreamCompression>();
         break;
     default:
-        return Status::InternalError(fmt::format("Unknown compress type: {}", type));
+        return Status::InternalError(fmt::format("Unknown compress type: {}", static_cast<int>(type)));
     }
 
     Status st = Status::OK();
